validação exclusão de tweet

+++ src/useCases/deleteTweet/DeleteTweetUseCase.ts
@@ -2,7 +2,7 @@ import { response } from 'express'
 import { prisma } from '../../lib/prisma'
 
 class DeleteTweetUseCase {
-  async execute(id: string) {
+  async execute(id: string, user_id: string) {
     const tweet = await prisma.tweet.findFirst({
       where: {
         id
@@ -11,6 +11,10 @@ class DeleteTweetUseCase {
 
     if (!tweet) {
       throw new Error('Tweet does not exist!')
+    }
+
+    if (tweet.user_id !== user_id) {
+      throw new Error('Only the author can remove the tweet!')
     }
 
     await prisma.tweet.delete({
